RiscVPlatformSecLib: Fix for 64-bit temporary ram address
The temporary ram base address can be greater than 32 bit. Fix it.

+++ Platform/RISC-V/PlatformPkg/Include/Library/RiscVPlatformSecLib.h
@@ -30,7 +30,7 @@ EFIAPI
 SecStartupPlatform (
   IN  UINTN   BootHartId,
   IN  VOID    *DeviceTreeAddress,
-  IN  UINT32  TempRamBase,
+  IN  UINT64  TempRamBase,
   IN  UINT32  TempRamSize
   );
